+++ Moq.Tests/Matchers/AnyMatcherFixture.cs
@@ -12,7 +12,6 @@ namespace Moq.Tests.Matchers
 			var expr = ToExpression<object>(() => It.IsAny<object>()).ToLambda().Body;
 
 			var matcher = MatcherFactory.CreateMatcher(expr, false);
-			matcher.Initialize(expr);
 
 			Assert.True(matcher.Matches(null));
 		}
@@ -23,7 +22,6 @@ namespace Moq.Tests.Matchers
 			var expr = ToExpression<object>(() => It.IsAny<object>()).ToLambda().Body;
 
 			var matcher = MatcherFactory.CreateMatcher(expr, false);
-			matcher.Initialize(expr);
 
 			Assert.True(matcher.Matches("foo"));
 		}
@@ -34,7 +32,6 @@ namespace Moq.Tests.Matchers
 			var expr = ToExpression<IDisposable>(() => It.IsAny<IDisposable>()).ToLambda().Body;
 
 			var matcher = MatcherFactory.CreateMatcher(expr, false);
-			matcher.Initialize(expr);
 
 			Assert.True(matcher.Matches(new Disposable()));
 		}
@@ -45,7 +42,6 @@ namespace Moq.Tests.Matchers
 			var expr = ToExpression<IFormatProvider>(() => It.IsAny<IFormatProvider>()).ToLambda().Body;
 
 			var matcher = MatcherFactory.CreateMatcher(expr, false);
-			matcher.Initialize(expr);
 
 			Assert.False(matcher.Matches("foo"));
 		}