+++ Source/IMatcher.cs
@@ -38,13 +38,10 @@
 //[This is the BSD license, see
 // http://www.opensource.org/licenses/bsd-license.php]
 
-using System.Linq.Expressions;
-
 namespace Moq
 {
 	internal interface IMatcher
 	{
-		void Initialize(Expression matcherExpression);
 		bool Matches(object value);
 	}
 }