+++ CHANGELOG.md
@@ -29,6 +29,10 @@ The format is loosely based on [Keep a Changelog](http://keepachangelog.com/en/1
 * Prevent Moq from relying on a mock's implementation of `IEnumerable<T>` (@stakx, #510)
 * Verification leaked internal `MockVerificationException` type; remove it (@stakx, #511)
 
+#### Obsoleted
+
+* `[Matcher]` has been deprecated in favor of `Match.Create` (@stakx, #514)
+
 
 ## 4.7.145 (2017-11-06)